add py packages
