+++ yummy-rs/yummy-delta/src/lib.rs
@@ -9,41 +9,10 @@ use actix_web::middleware::Logger;
 use actix_web::{web, App, HttpServer};
 use apply::DeltaApply;
 use delta::DeltaManager;
-use pyo3::prelude::*;
 use server::{
     append, create_table, details, health, list_stores, list_tables, optimize, overwrite,
     query_stream, vacuum,
 };
-
-#[pyfunction]
-fn run_apply(config_path: String) -> PyResult<String> {
-    tokio::runtime::Builder::new_current_thread()
-        .enable_all()
-        .build()
-        .unwrap()
-        .block_on(apply_delta(config_path))
-        .unwrap();
-    Ok("Ok".to_string())
-}
-
-#[pyfunction]
-fn run(config_path: String, host: String, port: u16, log_level: String) -> PyResult<String> {
-    tokio::runtime::Builder::new_current_thread()
-        .enable_all()
-        .build()
-        .unwrap()
-        .block_on(run_delta_server(config_path, host, port, log_level))
-        .unwrap();
-    Ok("Ok".to_string())
-}
-
-#[pymodule]
-fn yummy_delta(_py: Python<'_>, m: &PyModule) -> PyResult<()> {
-    m.add_function(wrap_pyfunction!(run, m)?)?;
-    m.add_function(wrap_pyfunction!(run_apply, m)?)?;
-
-    Ok(())
-}
 
 pub async fn apply_delta(config_path: String) -> std::io::Result<()> {
     DeltaApply::new(&config_path)
