+++ yummy-rs/yummy-features-py/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "yummy-features-py"
+version = "0.0.8"
+#version.workspace = true
+edition = "2021"
+repository = "https://github.com/yummyml/yummy"
+description = "Feature store server of the yummy-rs library"
+
+[lib]
+name = "yummy_features"
+crate-type = ["cdylib"]
+
+[build]
+rustflags = ["-C", "target-cpu=native"]
+
+[dependencies]
+yummy-features = { path = "../yummy-features" }
+tokio = "1.20.0"
+pyo3 = { version = "0.16.5", features = ["extension-module"] }
