+++ yummy-rs/yummy-features-py/src/lib.rs
@@ -1,0 +1,20 @@
+use pyo3::prelude::*;
+use yummy_features::serve_wrapper;
+
+#[pyfunction]
+fn serve(config_path: String, host: String, port: u16, log_level: String) -> PyResult<String> {
+    tokio::runtime::Builder::new_current_thread()
+        .enable_all()
+        .build()
+        .unwrap()
+        .block_on(serve_wrapper(config_path, host, port, log_level))
+        .unwrap();
+    Ok("Ok".to_string())
+}
+
+#[pymodule]
+fn yummy_features(_py: Python<'_>, m: &PyModule) -> PyResult<()> {
+    m.add_function(wrap_pyfunction!(serve, m)?)?;
+
+    Ok(())
+}
