+++ yummy-rs/yummy-mlflow-py/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "yummy-mlflow-py"
+version = "0.0.8"
+#version.workspace = true
+edition = "2021"
+repository = "https://github.com/yummyml/yummy"
+description = "MLflow serve of the yummy-rs library"
+
+[lib]
+name = "yummy_mlflow"
+crate-type = ["cdylib"]
+
+[build]
+rustflags = ["-C", "target-cpu=native"]
+
+[dependencies]
+yummy-mlflow = { path = "../yummy-mlflow" }
+tokio = "1.20.0"
+pyo3 = { version = "0.16.5", features = ["extension-module"] }
+
