+++ yummy-rs/yummy-mlflow/Cargo.toml
@@ -6,13 +6,6 @@ edition = "2021"
 repository = "https://github.com/yummyml/yummy"
 description = "MLflow serve of the yummy-rs library"
 
-[lib]
-name = "yummy_mlflow"
-crate-type = ["cdylib"]
-
-[build]
-rustflags = ["-C", "target-cpu=native"]
-
 [dependencies]
 actix-web = "4"
 serde = { version = "1.0", features = ["derive"] }
@@ -24,7 +17,6 @@ chrono = "0.4.22"
 derive_more = "0.99.17"
 #catboost = { git = "https://github.com/catboost/catboost", optional = true }
 lightgbm = "0.2.3"
-pyo3 = { version = "0.16.5", features = ["extension-module"] }
 thiserror = "1.0"
 #reqwest = "0.11.13"
