+++ yummy-rs/yummy-mlflow/src/lib.rs
@@ -7,26 +7,7 @@ use actix_web::middleware::Logger;
 use actix_web::{web, App, HttpServer};
 use config::MLConfig;
 use models::MLModelFactory;
-use pyo3::prelude::*;
 use server::{health, invocations};
-
-#[pyfunction]
-fn serve(model_path: String, host: String, port: u16, log_level: String) -> PyResult<String> {
-    tokio::runtime::Builder::new_current_thread()
-        .enable_all()
-        .build()
-        .unwrap()
-        .block_on(serve_mlflow_model(model_path, host, port, log_level))
-        .unwrap();
-    Ok("Ok".to_string())
-}
-
-#[pymodule]
-fn yummy_mlflow(_py: Python<'_>, m: &PyModule) -> PyResult<()> {
-    m.add_function(wrap_pyfunction!(serve, m)?)?;
-
-    Ok(())
-}
 
 pub async fn serve_mlflow_model(
     model_path: String,
